estate-sentry-api/requirements.txt
Django>=5.0.0,<5.2.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.3.0,<5.0.0
Pillow>=10.4.0
pytest>=8.3.0,<9.0.0
pytest-django>=4.7.0,<5.0.0
flake8>=7.0.0,<8.0.0
